refactor: move effectifs to actions
+ remove the components injection mechanism
