--- server/src/common/components/components.js
@@ -1,9 +1,0 @@
-import createEffectifs from "./effectifs.js";
-
-export default async (options = {}) => {
-  const effectifs = options.effectifs || createEffectifs();
-
-  return {
-    effectifs,
-  };
-};
